+++ packages/react-sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @phantom/react-sdk
 
+## 1.0.0-beta.17
+
+### Patch Changes
+
+- bd2219e: Fixed features access
+- Updated dependencies [bd2219e]
+  - @phantom/browser-sdk@1.0.0-beta.17
+
+## 1.0.0-beta.16
+
+### Patch Changes
+
+- bb3483b: Fix double connect issues , change provider bug
+- Updated dependencies [bb3483b]
+  - @phantom/browser-sdk@1.0.0-beta.16
+
 ## 1.0.0-beta.15
 
 ### Patch Changes